ml-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
sentence-transformers==3.1.1
faiss-cpu==1.8.0.post1
xgboost==2.1.1
lightgbm==4.5.0
scikit-learn==1.5.2
joblib==1.4.2
numpy==1.26.4
scipy==1.14.1
pandas==2.2.3
opencv-python==4.10.0.84
python-multipart==0.0.12
